[FEATURE] Supress adding HTTP debug header by extension configuration
The name of the route is sent as an additional header line in the HTTP
response to the client by default. For example:

X-Causal-Routing-Route: [ExtensionKey] Demo action with a parameter

Under certain circumstances is not desired.

This feature allows TYPO3 integrators to configure the extension and to
supress adding this header line. This can be done in the Extension Manager
(which requires admin privileges). The documentation has also been updated
to reflect this new feature.